move event model logic into query.js and add constants.js
