+++ assets/src/data/model/event/constants.js
@@ -1,0 +1,1 @@
+export const MODEL_NAME = 'event';
